Update README.md
Hanagaki Takemichi lives an unsatisfying life right up until his death. Waking up 12 years in the past, he reckons with the eventual fate of his friends and tries to prevent an unfortunate results in the future. Takemichi  is a freelancer that's reached the absolute pits of despair in his life.